formatting and fixes
