+++ components/CustomButton.js
@@ -1,10 +1,9 @@
-import { View, Text, StyleSheet } from 'react-native'
+import { Text, StyleSheet } from 'react-native'
 import React from 'react'
 import Pressable from 'react-native/Libraries/Components/Pressable/Pressable'
-import globalStyles from '../styles/globalStyles'
-import colors from '../styles/colors'
 import Ionicons from "@expo/vector-icons/Ionicons";
 import LoadingSpinner from './LoadingSpinner'
+import { colors, globalStyles } from '../styles';
 
 
 const CustomButton = ({containerStyles, textStyles, text, buttonConfig, disabled, icon, iconColor, loading, secondary}) => {
