+++ components/Posts/PostActionBar.js
@@ -1,7 +1,7 @@
-import { View, Text, Pressable } from "react-native";
+import { View, Pressable } from "react-native";
 import React from "react";
 import Ionicons from "@expo/vector-icons/Ionicons";
-import colors from "../../styles/colors";
+import { colors } from "../../styles";
 
 const PostActionBar = ({
     onPressZap,
