+++ features/homefeed/components/ReadMoreModal.js
@@ -1,11 +1,10 @@
 import { View, Text } from "react-native";
 import React from "react";
-import globalStyles from "../../../styles/globalStyles";
 import { ScrollView } from "react-native-gesture-handler";
 import CustomButton from "../../../components/CustomButton";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import colors from "../../../styles/colors";
 import { useParseContent } from "../../../hooks/useParseContent";
+import { globalStyles } from "../../../styles";
 
 
 const ReadMoreModal = ({ navigation,route }) => {
