+++ features/zaps/components/ZapItem.js
@@ -1,8 +1,8 @@
 import { useNavigation } from "@react-navigation/native";
 import { Pressable, View, Text } from "react-native";
+import Ionicons from "@expo/vector-icons/Ionicons";
 import colors from "../../../styles/colors";
 import globalStyles from "../../../styles/globalStyles";
-import Ionicons from "@expo/vector-icons/Ionicons";
 
 const getAge = (timestamp) => {
     const now = new Date();
